+++ server/web/src/index.tsx
@@ -7,6 +7,7 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { HashRouter as Router } from 'react-router';
 import './style/app-theme.less';
+import '@ant-design/v5-patch-for-react-19';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);