+++ server/web/src/pages/action-invocations/DeleteActionInvocation.tsx
@@ -1,6 +1,6 @@
 import { useDeleteActionInvocationMutation } from '../../api/actionInvocationsApi';
 import { apiNotification } from '../common/apiNotification';
-import { DeleteOutlined, DeleteTwoTone } from '@ant-design/icons';
+import { DeleteOutlined } from '@ant-design/icons';
 import { Button, Popconfirm, Tooltip } from 'antd';
 import { FC, ReactNode, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -37,7 +37,7 @@ const DeleteActionInvocation: FC<DeleteActionInvocationProps> = ({ id }): ReactN
 			cancelText={t('cancel')}
 			okButtonProps={{ icon: <DeleteOutlined />, type: 'primary', danger: true }}>
 			<Tooltip title={t('help_delete')} placement="right">
-				<Button loading={isLoading} key="del" icon={<DeleteTwoTone twoToneColor={'red'} />} type="text" danger>
+				<Button loading={isLoading} key="del" icon={<DeleteOutlined />} type="text" danger>
 					{t('delete')}
 				</Button>
 			</Tooltip>