+++ server/web/src/pages/webhooks/WebhooksPage.tsx
@@ -16,6 +16,8 @@ import { FC, useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
 
+const { Text, Title } = Typography;
+
 const DEFAULT_POLLING_INTERVAL = 10000;
 
 const WebhooksPage: FC = () => {
@@ -77,19 +79,19 @@ const WebhooksPage: FC = () => {
 		<>
 			<AppBreadcrumb items={[{ label: t('title'), active: true, path: '' }]} />
 			<PageHeader
-				className={'pl-0'}
+				className="pl-0"
 				title={
-					<Typography.Title level={4} ellipsis>
+					<Title level={4} ellipsis>
 						{t('title')}
 						<Tooltip placement="bottom" title={parse(t('help'))}>
 							<Button icon={<QuestionCircleOutlined />} type="link" />
 						</Tooltip>
-					</Typography.Title>
+					</Title>
 				}
 				extra={
 					<Space>
 						<Space>
-							<Typography.Text>{t('auto_refresh')}</Typography.Text>
+							<Text>{t('auto_refresh')}</Text>
 							<Switch
 								checkedChildren={t('on')}
 								unCheckedChildren={t('off')}
@@ -97,10 +99,10 @@ const WebhooksPage: FC = () => {
 								value={pollingInterval > 0}
 							/>
 						</Space>
-						<Tooltip title={t('reload_tooltip')} placement={'bottom'}>
+						<Tooltip title={t('reload_tooltip')} placement="bottom">
 							<Button
 								icon={<ReloadOutlined />}
-								type={'link'}
+								type="link"
 								onClick={invokeReload}
 								loading={isFetching}
 								disabled={isFetching || isLoading}